refactor(teams): split message bus helpers [OPE-349]
